style: extracted OnlineStatistics to its own component
