+++ src/components/Modal/Summary/Summary.tsx
@@ -1,19 +1,15 @@
-import { ShareIcon, CogIcon } from '@heroicons/react/solid'
+import { ShareIcon } from '@heroicons/react/solid'
 import ModalContainer, { ContainerProps } from '../ModalContainer'
 import HistogramChart from './HistogramChart/HistogramChart'
 import React, { memo, useEffect, useState } from 'react'
 import { GameStatistics } from '../../../lib/stats/types'
 import StatisticWindow from './StatisticWindow'
 import { getTimeLeft } from '../../../lib/time'
-import { WordFrequency } from '../../../lib/collection/types'
-import {
-  getThreeMostFrequentWords,
-  getWordsFrequency,
-} from '../../../lib/collection/collection'
+import OnlineStatistics from './OnlineStatistics'
 
 interface DataProps {
   gameStatistics: GameStatistics
-  shouldShowShareButton: boolean
+  isGameFinished: boolean
   submittedWords: string[]
 }
 
@@ -27,17 +23,12 @@ const Summary = ({
   shouldShow,
   onHide,
   gameStatistics,
-  shouldShowShareButton,
+  isGameFinished,
   submittedWords,
   onShare,
 }: Props) => {
   const [isCopied, setIsCopied] = useState(false)
   const [timeLeft, setTimeLeft] = useState(() => getTimeLeft())
-  const [isLoading, setIsLoading] = useState(false)
-  const [threeMostFrequentWord, setThreeMostFrequentWord] = useState<
-    WordFrequency[]
-  >([])
-  const [mostGuessedWord, setMostGuessedWord] = useState<WordFrequency>()
 
   const handleOnShare = (event: React.MouseEvent<HTMLButtonElement>) => {
     setIsCopied(true)
@@ -49,58 +40,11 @@ const Summary = ({
   }
 
   useEffect(() => {
-    setIsLoading(true)
     const interval = setInterval(() => {
       setTimeLeft(getTimeLeft())
     }, 1000)
     return () => clearInterval(interval)
   }, [])
-
-  const fetchOnlineStatistics = async () => {
-    setIsLoading(true)
-    const threeMostFrequentWords = await getThreeMostFrequentWords()
-    const mostGuessedWords = await getWordsFrequency(submittedWords)
-    setThreeMostFrequentWord(threeMostFrequentWords)
-    setMostGuessedWord(mostGuessedWords[0])
-    setIsLoading(false)
-  }
-
-  useEffect(() => {
-    if (!shouldShowShareButton) return
-    fetchOnlineStatistics()
-  }, [shouldShowShareButton])
-
-  const onlineStatistics =
-    !isLoading && threeMostFrequentWord.length > 0 && mostGuessedWord ? (
-      <>
-        <div>
-          <h1>คุณทายคำว่า</h1>
-          <div className="flex justify-center items-end">
-            <h1 className="text-xl">{mostGuessedWord?.word}</h1>
-          </div>
-          <h1>เหมือนกับคนอีก {mostGuessedWord?.frequency} คน</h1>
-        </div>
-        <div>
-          <h1>คำที่มีคนทายเยอะที่สุดในวันนี้</h1>
-          <div className="flex justify-center items-end">
-            <div className="order-2">
-              <span>{threeMostFrequentWord[0].word}</span>
-              <div className="h-10 w-16 bg-yellow-400"></div>
-            </div>
-            <div>
-              <span>{threeMostFrequentWord[1].word}</span>
-              <div className="h-6 w-16 bg-yellow-400"></div>
-            </div>
-            <div className="order-3">
-              <span>{threeMostFrequentWord[2].word}</span>
-              <div className="h-4 w-16 bg-yellow-400"></div>
-            </div>
-          </div>
-        </div>
-      </>
-    ) : (
-      <CogIcon className="animate-spin text-gray-400 w-8 h-8" />
-    )
 
   return (
     <ModalContainer shouldShow={shouldShow} onHide={onHide}>
@@ -127,9 +71,12 @@ const Summary = ({
         <div>
           <HistogramChart histogram={gameStatistics.histogram} />
         </div>
-        {shouldShowShareButton && (
+        {isGameFinished && (
           <div className="flex flex-col gap-6 items-center">
-            {onlineStatistics}
+            <OnlineStatistics
+              submittedWords={submittedWords}
+              isGameFinished={isGameFinished}
+            />
             <div>
               <button
                 className="p-4 bg-blue-500 hover:bg-blue-600 rounded text-white"
